Update README.md
For this problem, your should use the data in data folder. You should train your trees
on the examples in the training file data/a1a.train and evaluate on the test examples
in data/a1a.test. Remember that you should not look at or use your testing file until
your training is complete.
This is ID3 algorithm and it shows the error and depth restriction